Feature/operacao pedido

+++ src/main/java/br/com/pedido/models/Pedido_Produto.java
@@ -36,6 +36,6 @@ public class Pedido_Produto {
     private BigDecimal valorProduto;
 
     @Column(name = "total_produto")
-    private Double totalProduto;
+    private BigDecimal totalProduto;
 
 }
